Decoupling: Move PostLock and StatusCheck to web-story-editor
