+++ packages/story-editor/src/app/config/index.js
@@ -16,3 +16,4 @@
 
 export { default as ConfigProvider } from './configProvider';
 export { default as useConfig } from './useConfig';
+export { default as ConfigContext } from './context';
