+++ packages/story-editor/src/index.js
@@ -23,7 +23,10 @@ import Dialog from './components/dialog';
 export * from './components/transform';
 export * from './app/config';
 export * from './app/story';
+export * from './app/api';
+export * from './app/currentUser';
 export * from './components/previewPage';
+export * from './output';
 export * from './testUtils';
 
 export { default as base64Encode } from './utils/base64Encode';
