+++ packages/story-editor/src/storyEditor.js
@@ -42,8 +42,6 @@ import AutoSaveHandler from './components/autoSaveHandler';
 import { TransformProvider } from './components/transform';
 import { DropTargetsProvider } from './components/dropTargets';
 import { HelpCenterProvider } from './app/helpCenter';
-import StatusCheck from './components/statusCheck';
-import PostLock from './components/postLock';
 import Layout from './components/layout';
 import DevTools from './components/devTools';
 import { GlobalStyle as DefaultMoveableGlobalStyle } from './components/moveable/moveStyle';
@@ -60,7 +58,6 @@ function StoryEditor({ config, children }) {
         <ErrorBoundary>
           <ConfigProvider config={config}>
             <APIProvider>
-              <StatusCheck />
               <FileProvider>
                 <Media3pApiProvider>
                   <HistoryProvider size={50}>
@@ -68,7 +65,6 @@ function StoryEditor({ config, children }) {
                       <MetaBoxesProvider>
                         <StoryProvider storyId={storyId}>
                           <CurrentUserProvider>
-                            <PostLock />
                             <FontProvider>
                               <MediaProvider>
                                 <AutoSaveHandler />
