+++ packages/wp-story-editor/package.json
@@ -17,9 +17,12 @@
         "@web-stories-wp/react": "*",
         "@web-stories-wp/story-editor": "*",
         "@web-stories-wp/tracking": "*",
-        "flagged": "^2.0.1"
+        "flagged": "^2.0.1",
+	    "styled-components": "^5.3.1"
     },
     "devDependencies": {
-        "@testing-library/react": "^12.0.0"
+	    "@storybook/addon-actions": "^6.3.6",
+        "@testing-library/react": "^12.0.0",
+	    "prop-types": "^15.7.2"
     }
 }
