+++ packages/wp-story-editor/src/components/index.js
@@ -14,3 +14,5 @@
  * limitations under the License.
  */
 export { default as PostPublishDialog } from './postPublishDialog';
+export { default as PostLock } from './postLock';
+export { default as StatusCheck } from './statusCheck';
