+++ packages/wp-story-editor/src/components/postLock/index.js
