+++ packages/wp-story-editor/src/index.js
@@ -21,6 +21,7 @@
 // That's why the public path assignment is in its own dedicated module and imported here at the very top.
 // See https://webpack.js.org/guides/public-path/#on-the-fly
 import './publicPath';
+import './style.css'; // This way the general editor styles are loaded before all the component styles.
 
 /**
  * External dependencies
@@ -35,8 +36,7 @@ import { initializeTracking } from '@web-stories-wp/tracking';
 /**
  * Internal dependencies
  */
-import './style.css'; // This way the general editor styles are loaded before all the component styles.
-import { PostPublishDialog } from './components';
+import { PostPublishDialog, StatusCheck, PostLock } from './components';
 
 /**
  * Initializes the web stories editor.
@@ -60,6 +60,8 @@ const initialize = (id, config, flags) => {
       <StrictMode>
         <StoryEditor config={config}>
           <PostPublishDialog />
+          <StatusCheck />
+          <PostLock />
         </StoryEditor>
       </StrictMode>
     </FlagsProvider>,
